requirements.txt
mysql-connector-python
SQLAlchemy
pandas
pyyaml
flask
requests
Flask-APScheduler
python-dotenv